# core/openapi.yaml
openapi: 3.1.0
info:
  title: Madang core API
  version: 2.0.0
  summary: madang-core의 로컬 HTTP·WebSocket API.
  description: |
    데스크톱 앱은 이 API로 madang-core와 통신한다. 페이지 기록은 프로젝트
    폴더의 `.madang/`에, 전역 설정(config.yaml)과 profile.md는 앱 홈(`~/.madang`)에 있다.
    둘 다 core가 쓰며 앱은 자체 문서를 두지 않는다. core는 기록을 git에
    커밋하지 않는다(커밋 여부는 사용자가 정한다).

    규약:
    - 시각은 오프셋이 붙은 ISO 8601이다(`2026-09-24T09:40:00+09:00`).
    - 페이지 id는 `YYYY-MM-DD-<slug>`, 블록 id는 `bNN`이다. run 번호는
      페이지별로 늘어나며 재사용하지 않는다.
    - 페이지 안의 경로는 페이지 기준이다(`blocks/b05-base.json`). 프로젝트
      폴더 기준 경로에는 `repo:` 접두어가 붙는다
      (`repo:src/session/lock.ts`).
    - 오류는 `ApiError`로 돌려준다. 파일 검사에 실패하면 400과 함께
      `ValidationFailure`와 그 `issues`를 돌려준다.
    - 변경은 `/events` WebSocket으로 알린다. 앱은 재연결 후 전부 다시
      가져온다.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:7470
    description: 로컬 core. 포트는 다를 수 있으며 core가 core.port에 적는다.
...
paths:
...
  /pages/{page}:
    parameters:
      - $ref: "#/components/parameters/Page"
    get:
      tags: [pages]
      operationId: getPage
      summary: page.md와 블록 머리부.
      responses:
        "200":
          description: 페이지.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [pages]
      operationId: updatePage
      summary: 제목, 고정, 태그, 블록 순서를 바꾸거나 페이지를 다른 프로젝트로 옮긴다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageUpdate"
      responses:
        "200":
          description: 수정된 카드.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageCard"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [pages]
      operationId: deletePage
      summary: 페이지 폴더를 프로젝트의 `.madang/trash/`로 옮긴다. 휴지통에서 복원할 수 있다.
      responses:
        "204":
          description: 휴지통으로 옮겼다.
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /pages/{page}/blocks:
    parameters:
      - $ref: "#/components/parameters/Page"
    post:
      tags: [blocks]
      operationId: createBlock
      summary: doc, data, view 블록을 만들어 페이지 끝에 붙인다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockCreate"
      responses:
        "201":
          description: 새 블록 머리부.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockHeader"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
  /pages/{page}/blocks/{block}:
    parameters:
      - $ref: "#/components/parameters/Page"
      - $ref: "#/components/parameters/Block"
    get:
      tags: [blocks]
      operationId: getBlock
      summary: 블록 파일 전체.
      responses:
        "200":
          description: 머리부와 파일 내용.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [blocks]
      operationId: replaceBlock
      summary: 블록 파일 전체를 바꾼다(앱에서 직접 편집).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockContent"
      responses:
        "200":
          description: 저장된 블록.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [blocks]
      operationId: updateBlockHeader
      summary: 머리부 키만 바꾼다(bindings, presets, theme, title).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockHeaderPatch"
      responses:
        "200":
          description: 새 머리부.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockHeader"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [blocks]
      operationId: deleteBlock
      summary: 블록 파일을 `.madang/trash/`로 옮기고 page.md 순서에서 뺀다.
      responses:
        "204":
          description: 휴지통으로 옮겼다.
        "404":
          $ref: "#/components/responses/NotFound"
  /pages/{page}/memory:
    parameters:
      - $ref: "#/components/parameters/Page"
    get:
      tags: [memory]
      operationId: getMemory
      summary: 에이전트가 매 호출마다 읽는 3층 memory.
      responses:
        "200":
          description: profile.md(앱 홈), brief.md(`.madang/`), ledger.md(페이지).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "404":
          $ref: "#/components/responses/NotFound"
  /pages/{page}/memory/{layer}:
    parameters:
      - $ref: "#/components/parameters/Page"
      - name: layer
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/MemoryLayer"
    put:
      tags: [memory]
      operationId: saveMemory
      summary: memory 파일 하나를 저장한다. 검사를 통과해야 저장된다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryContent"
      responses:
        "200":
          description: 저장된 파일.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryFile"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /viewers:
    get:
      tags: [viewers]
      operationId: listViewers
      summary: 등록된 뷰어와 상태(끊김 포함).
      description: |
        등록부(`~/.madang/viewers.yaml`)의 항목은 정본 폴더를 가리키는
        참조다. 정본 폴더가 없거나 규격이 깨졌으면 `status`가 `broken`
        (끊김)이며, 그 뷰어를 쓰는 문서는 데이터를 표로 대신 보여 준다.
        `project`를 주면 그 프로젝트 `.madang/viewers/`의 뷰어가 먼저 온다
        (같은 이름이면 그 프로젝트 안에서 우선).
      parameters:
        - name: project
          in: query
          required: false
          description: 프로젝트 전용 뷰어도 넣을 프로젝트 id.
          schema:
            type: string
          example: jobs
      responses:
        "200":
          description: 뷰어 목록.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ViewerStatus"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    post:
      tags: [viewers]
      operationId: registerViewer
      summary: 뷰어 폴더를 `viewer.json`의 이름으로 등록한다.
      description: |
        복사하지 않고 폴더를 참조한다. `pinned`면 지금 내용의 해시로
        `cache/`에 사본을 만든다. 규격이 틀리면 400, 같은 이름이 이미 있으면
        409. 등록하면 `viewer.changed`를 낸다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViewerRegister"
      responses:
        "201":
          description: 등록한 뷰어.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewerStatus"
        "400":
          $ref: "#/components/responses/Invalid"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  parameters:
...
    Page:
      name: page
      in: path
      required: true
      description: 페이지 id.
      schema:
        type: string
      example: 2026-09-24-resume
    Block:
      name: block
      in: path
      required: true
      description: 블록 id.
      schema:
        type: string
      example: b04
...
  responses:
    Invalid:
      description: 요청 또는 결과 파일이 검사에 실패했다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationFailure"
    NotFound:
      description: 프로젝트, 페이지, 블록, run, 파일이 없다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: not_found
            message: page '2026-09-30-missing' not found
    Conflict:
      description: 현재 상태에서는 변경할 수 없다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: conflict
            message: "project 'jobs' still has child projects: jobs-2025"
  schemas:
...
    PageStatus:
      type: string
      enum: [planning, doing, blocked, review, done]
...
    PageCard:
      type: object
      description: 페이지 목록의 카드 하나.
      required: [id, project, title, status, pinned, tags, block_counts]
      properties:
        id:
          type: string
        project:
          type: string
          description: 페이지가 속한 프로젝트 id.
        title:
          type: string
        kind:
          type: string
          description: 기본 요청 종류(router 힌트).
        status:
          $ref: "#/components/schemas/PageStatus"
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        pinned:
          type: boolean
        tags:
          type: array
          items:
            type: string
        preview:
          type: [string, "null"]
          description: 마지막 메시지 한 줄.
        block_counts:
          type: object
          description: 블록 종류별 개수. run은 `run`으로 센다.
          additionalProperties:
            type: integer
        last_run:
          $ref: "#/components/schemas/RunRef"
          description: 페이지에 run이 아직 없으면 없다.
        first_view:
          type: [string, "null"]
          description: 카드 썸네일용 첫 view 블록의 id.
      examples:
        - id: 2026-09-24-resume
          project: jobs
          title: 이력서
          kind: build
          status: review
          created: "2026-09-24T08:10:00+09:00"
          updated: "2026-09-24T09:31:41+09:00"
          pinned: true
          tags: [이력서]
          preview: work 순서를 바꿨습니다. 펄어비스 경력이 맨 위입니다.
          block_counts: {message: 9, view: 1, data: 2, run: 2}
          last_run: {n: 2, runner: codex, model: gpt-6-luna, result_status: review}
          first_view: b04
...
    PageUpdate:
      type: object
      properties:
        title:
          type: string
        pinned:
          type: boolean
        tags:
          type: array
          items:
            type: string
        blocks_order:
          type: array
          description: page.md `blocks`의 새 순서 전체.
          items:
            type: string
        project:
          type: string
          description: 페이지 폴더를 옮길 프로젝트 id(그 프로젝트의 `.madang/pages/`로 옮긴다).
      examples:
        - pinned: true
          tags: [이력서]
        - project: archive
    PageDetail:
      type: object
      description: page.md, 본문 순서대로의 블록 머리부, run 목록.
      required: [id, project, title, status, pinned, tags, blocks, runs, unknown_files]
      properties:
        id:
          type: string
        project:
          type: string
          description: 페이지가 속한 프로젝트 id.
        title:
          type: string
        kind:
          type: string
        status:
          $ref: "#/components/schemas/PageStatus"
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        pinned:
          type: boolean
        tags:
          type: array
          items:
            type: string
        overview:
          type: string
          description: page.md의 Markdown 본문.
        blocks:
          type: array
          description: page.md `blocks` 순서의 블록 머리부.
          items:
            $ref: "#/components/schemas/BlockHeader"
        runs:
          type: array
          description: >-
            run 요약, 오름차순. 각 run은 `trigger.message`의 메시지 뒤에
            보여 준다.
          items:
            $ref: "#/components/schemas/RunRecord"
        unknown_files:
          type: array
          description: 미처리 알 수 없는 파일. 비어 있지 않으면 배너를 보여 준다.
          items:
            $ref: "#/components/schemas/UnknownFile"
        waiting:
          $ref: "#/components/schemas/FlowWaitingData"
          description: >-
            flow가 사람 결정이나 runner를 기다리는 동안에만 있다.
      examples:
        - id: 2026-09-24-resume
          project: jobs
          title: 이력서
          kind: build
          status: review
          created: "2026-09-24T08:10:00+09:00"
          updated: "2026-09-24T09:31:41+09:00"
          pinned: true
          tags: [이력서]
          overview: ""
          blocks:
            - {id: b01, type: message, role: user, ts: "2026-09-24T08:11:02+09:00", text: 이력서 템플릿 만들어줘, target: {}}
            - {id: b02, type: message, role: router, ts: "2026-09-24T08:11:03+09:00", run: 1, text: "kind=design conf=0.91 → claude/claude-opus-5-5 high"}
            - {id: b03, type: message, role: agent, ts: "2026-09-24T08:14:40+09:00", run: 1, text: "이력서 뷰(b04)와 기본 데이터(b05), 펄어비스용 덮어쓰기(b06)를 만들었습니다."}
            - id: b04
              type: view
              file: blocks/b04-resume.view.md
              title: 이력서
              created_by: run 1
              template: resume@1
              bindings: {base: b05, overlay: b06}
              presets:
                기본: {base: b05}
                펄어비스: {base: b05, overlay: b06}
              active_preset: 펄어비스
              theme: {color.accent: "#3a5bd9"}
            - {id: b05, type: data, file: blocks/b05-base.json, title: 기본 데이터, created_by: run 1, format: json}
            - {id: b06, type: data, file: blocks/b06-pearlabyss.json, title: 펄어비스 덮어쓰기, created_by: run 1, format: json}
            - {id: b07, type: message, role: user, ts: "2026-09-24T09:30:58+09:00", text: 펄어비스 경력을 맨 위로, target: {block: b04, elements: ["work[1]"], mode: edit}}
            - {id: b08, type: message, role: router, ts: "2026-09-24T09:31:00+09:00", run: 2, text: "kind=small conf=1.00 → codex/gpt-6-luna high"}
            - {id: b09, type: message, role: agent, ts: "2026-09-24T09:31:41+09:00", run: 2, text: work 순서를 바꿨습니다. 펄어비스 경력이 맨 위입니다.}
          runs:
            - n: 1
              started: "2026-09-24T08:11:03+09:00"
              finished: "2026-09-24T08:14:40+09:00"
              trigger: {message: b01}
              kind: design
              tier: 1
              runner: claude
              model: claude-opus-5-5
              effort: high
              input:
                parts: {system_est: 24600, profile: 110, brief: 1840, ledger: 380, contract: 420, target: 0, request: 12}
                total_est: 27362
              usage: {input: 27410, cached: 0, output: 3120}
              changed_files: [blocks/b04-resume.view.md, blocks/b05-base.json, blocks/b06-pearlabyss.json, ledger.md]
              unknown_files: []
              verify: {cmd: null, ok: null}
              result_status: review
              events_log: runs/1.jsonl
            - n: 2
              started: "2026-09-24T09:31:00+09:00"
              finished: "2026-09-24T09:31:41+09:00"
              trigger: {message: b07, target: b04, elements: ["work[1]"], mode: edit}
              kind: small
              tier: 1
              runner: codex
              model: gpt-6-luna
              effort: high
              input:
                parts: {system_est: 24600, profile: 110, brief: 1840, ledger: 1320, contract: 420, target: 900, request: 40}
                total_est: 29230
              usage: {input: 29104, cached: 24310, output: 612}
              changed_files: [blocks/b05-base.json]
              unknown_files: []
              verify: {cmd: null, ok: null}
              result_status: review
              events_log: runs/2.jsonl
          unknown_files: []

    # 블록
...
    Block:
      type: object
      description: 블록 머리부와 파일 전체.
      required: [header, content]
      properties:
        header:
          $ref: "#/components/schemas/BlockHeader"
        content:
          type: string
          description: >-
            저장된 그대로의 파일 텍스트. 메시지는 메시지 본문, 데이터 블록은
            JSON 또는 CSV 텍스트.
      examples:
        - header: {id: b05, type: data, file: blocks/b05-base.json, title: 기본 데이터, created_by: run 1, format: json, schema: "resume#/base"}
          content: |
            {
              "name": "김마당",
              "title": "백엔드 엔지니어",
              "contact": {"email": "madang@example.com", "location": "서울"},
              "summary": "대규모 트래픽 서비스의 인증·세션 계층을 설계하고 운영해 온 8년 차 백엔드 엔지니어입니다.",
              "work": [
                {"company": "펄어비스", "role": "게임 서버 엔지니어", "period": "2019.01 – 2022.02", "highlights": ["동시 접속 30만 매치메이킹 서버 개발"]},
                {"company": "한빛소프트", "role": "시니어 백엔드 엔지니어", "period": "2022.03 – 현재", "highlights": ["세션 갱신 경합 제거로 로그인 오류율 92% 감소"]}
              ],
              "skills": ["Kotlin", "TypeScript", "PostgreSQL", "Redis"]
            }
    BlockContent:
      type: object
      required: [content]
      properties:
        content:
          type: string
      examples:
        - content: |
            {"title": "게임 서버 엔지니어 (펄어비스 지원용)"}
...
    BlockCreate:
      type: object
      description: >-
        `doc`과 `data`는 `content`를, `view`는 `template`과 `bindings`나
        `data`를 받는다.
      required: [type, name]
      properties:
        type:
          type: string
          enum: [doc, data, view]
        name:
          type: string
          description: 파일 이름용 slug(`bNN-<name>.<ext>`).
        content:
          type: string
        format:
          type: string
          enum: [json, csv]
          description: 데이터 형식. 기본값 json.
        template:
          type: string
          description: "`name` 또는 `name@version`."
        bindings:
          type: object
          additionalProperties:
            type: string
        data:
          type: array
          items:
            type: string
          description: >-
            `view`의 데이터 블록. `bNN`은 비어 있는 슬롯에 순서대로,
            `slot=bNN`은 그 슬롯에 묶는다. `bindings`와 함께 쓸 수 있다.
        title:
          type: string
          description: "`view`의 제목."
        in_run:
          type: boolean
          default: false
          description: >-
            에이전트 실행 안에서 보낸 요청이면 참(`madang view create`가
            `MADANG_PAGE`로 불렸을 때). 참이고 그 페이지에서 실행이 돌고
            있을 때만 `created_by`에 실행 번호를 붙인다.
      examples:
        - type: view
          name: resume
          template: resume@1
          bindings: {base: b05}

    # 메모리
    MemoryLayer:
      type: string
      description: >-
        메모리 층. profile은 앱 홈 profile.md(나), brief는 프로젝트
        `.madang/brief.md`, ledger는 페이지 ledger.md.
      enum: [profile, brief, ledger]
...
    MemoryContent:
      type: object
      required: [content]
      properties:
        content:
          type: string
      examples:
        - content: |
            # 나에 대해
            한국어로 답한다.

    # 메시지와 입력
...